==> notebook/2_feature_engineering.ipynb <==
import pandas as pd
import numpy as np
import joblib
from sklearn.impute import SimpleImputer
from sklearn.preprocessing import StandardScaler, MinMaxScaler, OneHotEncoder

house_variables = joblib.load('output/x_train.pkl')
house_target = joblib.load('output/y_train.pkl')

house_variables.head().T

house_variables.isnull().sum().sort_values(ascending=False) / len(house_variables)

NUM = ["MSSubClass", "LotFrontage", "LotArea", "YearBuilt", "YearRemodAdd", "MasVnrArea", "BsmtFinSF2",
       "BsmtUnfSF", "TotalBsmtSF", "1stFlrSF", "2ndFlrSF", "GrLivArea", "FullBath", "HalfBath", "BedroomAbvGr", "KitchenAbvGr",
      "TotRmsAbvGrd", "Fireplaces", "GarageYrBlt", "GarageCars", "GarageArea", "WoodDeckSF", "OpenPorchSF", "EnclosedPorch", "3SsnPorch", "ScreenPorch", "PoolArea", "MiscVal", "MoSold", "YrSold"]

CAT = ['MSZoning', 'Street', 'Alley', 'LotShape', 'LandContour', 'Utilities',
       'LotConfig', 'LandSlope', 'Neighborhood', 'Condition1', 'Condition2',
       'BldgType', 'HouseStyle', 'OverallQual', 'OverallCond', 'RoofStyle',
       'RoofMatl', 'Exterior1st', 'Exterior2nd', 'MasVnrType', 'ExterQual',
       'ExterCond', 'Foundation', 'BsmtQual', 'BsmtCond', 'BsmtExposure',
       'BsmtFinType1', 'BsmtFinSF1', 'BsmtFinType2', 'Heating', 'HeatingQC',
       'CentralAir', 'Electrical', 'LowQualFinSF', 'BsmtFullBath',
       'BsmtHalfBath', 'KitchenQual', 'Functional', 'FireplaceQu',
       'GarageType', 'GarageFinish', 'GarageQual', 'GarageCond', 'PavedDrive',
       'PoolQC', 'Fence', 'MiscFeature', 'SaleType', 'SaleCondition']

house_numerical = house_variables[NUM]
house_categorical = house_variables.drop(NUM, axis=1)

house_numerical

house_categorical

house_numerical.isnull().sum().sort_values(ascending=False) / len(house_numerical)

house_categorical.isnull().sum().sort_values(ascending=False) / len(house_categorical)

## IMPUTING DATA

def numerical_imputer(numerical,
                    state = 'transform'):
    
    index = numerical.index
    cols = numerical.columns
    
    if state == 'fit':
        imputer = SimpleImputer(
            missing_values=np.nan,
            strategy="mean")

        imputer.fit(numerical)
        joblib.dump(imputer,
                    "output/numerical_imputer.pkl")
    elif state == 'transform':
        imputer = joblib.load("output/numerical_imputer.pkl")
        
    imputed = imputer.transform(numerical)
    imputed = pd.DataFrame(imputed)
    imputed.index = index
    imputed.columns = cols
    return imputed

def categorical_imputer(df_categorical):
    df = df_categorical.copy()
    df.fillna(value = 'KOSONG', inplace=True)
    return df

df_numerical_imputed = numerical_imputer(house_numerical, state='fit')
df_categorical_imputed = categorical_imputer(house_categorical)

df_categorical_imputed.head()

# One Hot Encoding

def one_hot_encoder(x_cat,
                    state='fit'):
    df = x_cat.copy()
    index = x_cat.index
    col = x_cat.columns
    
    if state == 'fit':
        encoder = OneHotEncoder(sparse=False,handle_unknown='ignore')
        encoder.fit(x_cat)
        joblib.dump(encoder,
                    "output/onehotencoder.pkl")
        
    elif state == 'transform':
        encoder = joblib.load("output/onehotencoder.pkl")
    
    encoded = encoder.transform(x_cat)
    feat_names = encoder.get_feature_names_out(col)
    encoded = pd.DataFrame(encoded)
    encoded.index = index
    encoded.columns = feat_names
    return encoded

df_categorical_encoded = one_hot_encoder(df_categorical_imputed)

df_categorical_encoded.head()

## Join Data

df_joined = pd.concat([df_numerical_imputed, df_categorical_encoded], axis=1)

import sklearn
sklearn.__version__

df_joined

# STANDARDIZE ALL

def normalization(x_all,
                  state = 'fit'):
    index = x_all.index
    cols = x_all.columns
    

    if state == 'fit':
        normalizer = StandardScaler()
        normalizer.fit(x_all)
        joblib.dump(normalizer,
                    "output/normalizer.pkl")

    elif state == 'transform':
        normalizer = joblib.load("output/normalizer.pkl")
        
    normalized = normalizer.transform(x_all)
    normalized = pd.DataFrame(normalized)
    normalized.index = index
    normalized.columns = cols
    return normalized

df_normalized = normalization(df_joined)

df_normalized

joblib.dump(df_normalized, 'output/preprocessed_x_train.pkl')

import yaml

f = open("src/params/preprocess_params.yaml", "r")
params = yaml.load(f, Loader=yaml.SafeLoader)
f.close()

params

yaml.__version__

def run(params, xpath, ypath, dump_path, state='fit'):
    house_variables = joblib.load(xpath)
    house_target = joblib.load(ypath)
    
    house_numerical = house_variables[params['NUM_COLUMN']]
    house_categorical = house_variables[params['CAT_COLUMN']]
    
    df_numerical_imputed = numerical_imputer(house_numerical, state=state)
    df_categorical_imputed = categorical_imputer(house_categorical)
    
    df_categorical_encoded = one_hot_encoder(df_categorical_imputed, state=state)
    
    df_joined = pd.concat([df_numerical_imputed, df_categorical_encoded], axis=1)
    
    df_normalized = normalization(df_joined, state=state)
    
    joblib.dump(df_normalized, dump_path)

run(params,'output/x_valid.pkl', 'output/y_valid.pkl', 'output/preprocessed_x_valid.pkl', state='transform')
